requirements.txt
transformers==4.5.1
jsonlines==2.0.0
scikit-learn==0.24.2
seqeval==1.2.2
sklearn-crfsuite==0.3.6
tensorboard==2.5.0
pytorch-crf==0.7.2
torch==1.8.1